+++ pkg/unittest/lib/test_controller.js
@@ -38,7 +38,7 @@ window.onerror = function (message, url, lineNumber) {
 // HTML Imports allows a document to link to other HTMLs documents via
 // <link rel=import>. It also allows for those other documents to contain
 // <script> tags, which must be run before scripts on the main page.
-// We have package:html_import to polyfill this feature, and it will handle
+// We have package:web_components to polyfill this feature, and it will handle
 // starting Dartium/content_shell in that case. HTML Imports is used by Polymer,
 // but it could be used by itself too. See the specification:
 // https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html