+++ tests/html/custom/js_custom_test.dart
@@ -28,7 +28,7 @@ main() {
 
   var registered = false;
   setUp(() {
-    return loadPolyfills().then((_) {
+    return customElementsReady.then((_) {
       if (!registered) {
         registered = true;
         document.register(A.tag, A);