+++ pkg/polymer/test/prop_attr_reflection_test.dart
@@ -11,23 +11,23 @@ import 'package:polymer/polymer.dart';
 class XFoo extends PolymerElement {
   XFoo.created() : super.created();
 
-  @observable var foo = '';
-  @observable String baz = '';
+  @published var foo = '';
+  @published String baz = '';
 }
 
 class XBar extends XFoo {
   XBar.created() : super.created();
 
-  @observable int zot = 3;
-  @observable bool zim = false;
-  @observable String str = 'str';
-  @observable Object obj;
+  @published int zot = 3;
+  @published bool zim = false;
+  @published String str = 'str';
+  @published Object obj;
 }
 
 class XCompose extends PolymerElement {
   XCompose.created() : super.created();
 
-  @observable bool zim = false;
+  @published bool zim = false;
 }
 
 Future onAttributeChange(Element node) {