+++ pkg/polymer/CHANGELOG.md
@@ -4,6 +4,10 @@ This file contains highlights of what changes on each version of the polymer
 package. We will also note important changes to the polyfill packages (observe,
 web_components, and template_binding) if they impact polymer.
 
+#### Pub version 0.10.0+1
+  * Better error message on failures in pub-serve/pub-build when pubspec.yaml
+    is missing or has a wrong configuration for the polymer transformers.
+
 #### Pub version 0.10.0
   * Interop with polymer-js elements now works.
   * Polymer polyfills are now consolidated in package:web_components, which is